+++ src/main/java/com/hexagonal/microservicio_plazoleta/domain/api/IDishesServicePort.java
@@ -7,7 +7,7 @@ public interface IDishesServicePort {
 
     void saveDishes(Dishes dishes);
 
-    void updateDish(Long dishId, Integer price, String description, Long ownerId);
+    void updateDish(Long dishId, Double price, String description, Long ownerId);
 
     void updateDishStatus(Long dishId, Boolean active, Long ownerId);
