+++ src/main/java/com/hexagonal/microservicio_plazoleta/domain/model/Dishes.java
@@ -10,7 +10,7 @@ public class Dishes {
 
     private String name;
 
-    private Integer price;
+    private Double price;
 
     private String description;
 
@@ -24,7 +24,7 @@ public class Dishes {
 
     private Long ownerId;
 
-    public Dishes(Long id, String name, Integer price, String description, String imageUrl, String category, Long restaurantId, Long ownerId) {
+    public Dishes(Long id, String name, Double price, String description, String imageUrl, String category, Long restaurantId, Long ownerId) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -51,11 +51,11 @@ public class Dishes {
         this.name = name;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
